% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackageclasswork1.R
\name{retrieve_answer}
\alias{retrieve_answer}
\title{R DATA ANALYSIS}
\usage{
retrieve_answer(exercise_num)
}
\arguments{
\item{exercise_num}{The number of the exercise (between 1 and 6).}
}
\value{
The solution for the specified exercise.
}
\description{
R DATA ANALYSIS
}
\examples{
retrieve_answer(2)
retrieve_answer(5)
}
